Separated the Task class into a Task interface and AbstractTask class.
